% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BJ.R
\name{BJ}
\alias{BJ}
\title{BJ.R}
\usage{
BJ(test_stats, pairwise_cors)
}
\arguments{
\item{test_stats}{Vector of all individual (first-level) test statistics}

\item{pairwise_cors}{A vector of all d(d-1)/2 pairwise correlations between the test 
statistics, where d is total number of test statistics in the set.}
}
\value{
A list with the elements:
\item{BJ}{The observed Berk-Jones test statistic.}
\item{BJ_pvalue}{The p-value of this observed value, given the size of the set and 
correlation structure.}
}
\description{
Given a vector of individual test statistics and their pairwise correlations, calculate 
the Berk-Jones second-level test statistic and it's p-value.
}
\examples{
BJ(test_stats=1:5, d=5, pairwise_cors=rep(0.3,10))
}

