% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_score_stats.R
\name{calc_score_stats}
\alias{calc_score_stats}
\title{calc_score_stats.R}
\usage{
calc_score_stats(null_model, factor_matrix, model_type)
}
\arguments{
\item{null_model}{A fitted R regression model}

\item{factor_matrix}{An n*p matrix with each factor as one column.  There should be no missing data.}

\item{model_type}{Either "linear" or "logistic"}
}
\value{
A list with the elements:
\item{test_stats}{The p score test statistics.}
\item{cor_mat}{The p*p matrix giving the pairwise correlation of every two test statistics.}
}
\description{
Starting with individual-level data on p factors, generate score test statistics for each
factor for input into GBJ/GHC/HC/BJ/minP.  Also get the correlations between these test statistics. 
Designed to be used with linear or logistic regression null models.
}
\examples{
Y <- rbinom(n=100, size=1, prob=0.5)
null_mod <- glm(Y~1, family=binomial(link="logit"))
factor_mat <- matrix(data=rnorm(n=100*5), nrow=100)
calc_score_stats(null_mod, factor_mat, "logistic")
}

