% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebb_loglik.R
\name{ebb_loglik}
\alias{ebb_loglik}
\title{ebb_loglik.R}
\usage{
ebb_loglik(x, d)
}
\arguments{
\item{x}{A vector of length 3 with (1) value of outcome (2) mu parameter (3) gamma parameter}

\item{d}{The number of test stsatistics in the set}
}
\value{
log( Pr(V=x[1]) ) where V~EBB(mu,gamma; d)
}
\description{
An internal function only (doesn't do error checking or take care of boundary cases).
The log-likelihood (log-PMF) calculation for the Extended Beta-Binomial proposed by Prentice (1986).
Takes in a vectorized argument because we apply() it in run_BB_GBJ().
}
\examples{
ebb_loglik(x=c(1, 0.5, 0.1), d=10)	
}
\keyword{internal}
