% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FGFR2}
\alias{FGFR2}
\title{Genotypes at FGFR2 SNPs for subjects from 'GBR' population in the 1000 Genomes Project.}
\format{A matrix with 91 rows (one for each subject) and 64 columns (one for each SNP)}
\source{
\url{http://www.internationalgenome.org/data}
}
\usage{
data(FGFR2)
}
\description{
A dataset containing the genotypes (number of minor alleles) for each of 91 subjects
from the 'GBR' population in the 1000 Genomes Projects.  There are 64 SNPs documented
here, all residing in the FGFR2 gene.
}
\keyword{datasets}
