% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbr_pcs}
\alias{gbr_pcs}
\title{Simulated Principal Components for 'GBR' population in the 1000 Genomes Project.}
\format{A matrix with 91 rows (one for each subject) and 5 columns (one for each PC)}
\usage{
data(gbr_pcs)
}
\description{
A dataset containing 5 simulated Principal Components (PCs) for each of 91 subjects
from the 'GBR' population in the 1000 Genomes Projects.  These would normally be
used as covariates in a regression model to control for population stratification.
}
\keyword{datasets}
