% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHC.R
\name{GHC}
\alias{GHC}
\title{GHC.R}
\usage{
GHC(test_stats, pairwise_cors)
}
\arguments{
\item{test_stats}{Vector of all individual (first-level) test statistics}

\item{pairwise_cors}{A vector of all d(d-1)/2 pairwise correlations between the test 
statistics, where d is total number of test statistics in the set.}
}
\value{
A list with the elements:
\item{GHC}{The observed Generalized Higher Criticism test statistic.}
\item{GHC_pvalue}{The p-value of this observed value, given the size of the set and 
correlation structure.}
}
\description{
Given a vector of individual test statistics and their pairwise correlations, calculate 
the Generalized Higher Criticism second-level test statistic and it's p-value.
}
\examples{
GHC(test_stats=1:5, d=5, pairwise_cors=rep(0.3,10))
}

