% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHC_objective.R
\name{GHC_objective}
\alias{GHC_objective}
\title{GHC_objective.R}
\usage{
GHC_objective(x, k, d, offset, pairwise_cors)
}
\arguments{
\item{x}{The p-value of the kth p-value order statistic.}

\item{k}{Which objective to use.}

\item{d}{The size of the set.}

\item{offset}{Used to zero the correct value when we put this into uniroot}

\item{pairwise_cors}{A vector of all d(d-1)/2 pairwise correlations between the test 
statistics, where d is total number of test statistics in the set.}
}
\value{
The value of the kth HC objective
}
\description{
Internal function to calculate the value of the kth GHC objective (possibly minus an offset)
when the kth p-value order statistic is x, for a set of size d.
}
\examples{
GHC_objective(x=0.1, k=2, d=5, offset=0, pairwise_cors=rep(0.2,10))
}
\keyword{internal}
