% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herm_poly_diff_t.R
\name{herm_poly_diff_t}
\alias{herm_poly_diff_t}
\title{herm_poly_diff_t.R}
\usage{
herm_poly_diff_t(t1, t2, pairwise_cors)
}
\arguments{
\item{t1}{t1 and t2 are interchangeable.  One should be t-mu and one should be -t-mu.}

\item{t2}{t1 and t2 are interchangeable.  One should be t-mu and one should be -t-mu.}

\item{pairwise_cors}{A vector of all d(d-1)/2 pairwise correlations between the test 
statistics, where d is total number of test statistics in the set.}
}
\value{
The result of the infinite sum.
}
\description{
Internal function to add up the infinite sum term in calculating the variance of S(t) when
the test statistics Z_1,...,Z_d have a non-zero mean.  See GBJ paper for more details.
}
\examples{
herm_poly_diff_t(t1=0, t2=0, pairwise_cors=rep(0.3, 10))
}
\keyword{internal}
