% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnorm_mu.R
\name{qnorm_mu}
\alias{qnorm_mu}
\title{qnorm_mu.R}
\usage{
qnorm_mu(mu, t, kkk, d)
}
\arguments{
\item{mu}{The mean of the normal random variable.}

\item{t}{The threshold (boundaries) we are interested in.}

\item{kkk}{We decided to make you input the fraction in two parts.}

\item{d}{We decided to make you input the fraction in two parts.}
}
\value{
Pr(-t < Z < t) - kkk/d for Z~N(mu,1).
}
\description{
Internal function to calculate Pr(-t < Z < t) - kkk/d for Z~N(mu,1).
Take the root of this function to find the mu such that P(|Z|>=t_k) = k/d.
}
\examples{
qnorm_mu(mu=0, t=1.96, kkk=1, d=5)		# Should return 0
}
\keyword{internal}
