% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{surv}
\alias{surv}
\title{surv.R}
\usage{
surv(x)
}
\arguments{
\item{x}{Vector of quantiles}
}
\value{
Probability that a standard normal random variable is greater than x.
}
\description{
Survival (1 minus the CDF) function of standard normal random variable.
}
\examples{
surv(0)		# Should return 0.5
}
