% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHC.R
\name{GHC}
\alias{GHC}
\title{GHC.R}
\usage{
GHC(test_stats, pairwise_cors)
}
\arguments{
\item{test_stats}{Vector of all individual (first-level) test statistics}

\item{pairwise_cors}{A vector of all d(d-1)/2 pairwise correlations between the test 
statistics, where d is total number of test statistics in the set.}
}
\value{
A list with the elements:
\item{GHC}{The observed Generalized Higher Criticism test statistic.}
\item{GHC_pvalue}{The p-value of this observed value, given the size of the set and 
correlation structure.}
}
\description{
Given a vector of individual test statistics and their pairwise correlations, calculate 
the Generalized Higher Criticism (GHC) second-level test statistic and it's p-value.
}
\examples{
# Should return statistic = 1.987733 and p_value = 0.2758522
set.seed(100)
test_stats <- rnorm(5) + rep(1,5)
GHC(test_stats, pairwise_cors=rep(0.2,10))
}

