% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC.R
\name{HC}
\alias{HC}
\title{HC.R}
\usage{
HC(test_stats, pairwise_cors)
}
\arguments{
\item{test_stats}{A scalar or vector of threshold points (magnitudes of test statistics)}

\item{pairwise_cors}{A vector of all d(d-1)/2 pairwise correlations between the test 
statistics, where d is total number of test statistics in the set.}
}
\value{
A list with the elements:
\item{HC}{The observed Higher Criticism test statistic.}
\item{HC_pvalue}{The p-value of this observed value, given the size of the set and 
correlation structure.}
}
\description{
Given a vector of individual test statistics and their pairwise correlations, calculate 
the Higher Criticism second-level test statistic and it's p-value.
}
\examples{
# Should return statistic = 2.067475 and p_value = 0.2755146
set.seed(100)
test_stats = rnorm(5) + rep(1,5)
HC(test_stats, pairwise_cors=rep(0.2,10))
}

